Remove functions that modify edit links
Since the lesson plans are no longer sourced from GitHub, the edit
links should not point to GitHub repositories.

This also eliminates the `WPOrg_Learn\Lesson_Plan` class, since with
this commit, everything in it has either been moved elsewhere or
removed.
